linux/tests/requirements.txt
pytest==7.1.1
google-cloud-compute==1.1.0